+++ src/models/instructions/msg_unknown.ts
@@ -1,0 +1,23 @@
+import * as R from 'ramda';
+
+// model to handle all unknown data
+class InstructionUnknown {
+  public type: string;
+  public json: JSON;
+
+  constructor(payload: any) {
+    this.type = payload.type;
+    this.json = payload.json;
+  }
+
+  static fromJson(json: any) {
+    return new InstructionUnknown({
+      type: R.pathOr('', ['type'], json),
+      program: R.pathOr('', ['program'], json),
+      rawData: R.pathOr('', ['raw_data'], json),
+      json,
+    });
+  }
+}
+
+export default InstructionUnknown;
