--- src/models/msg/types.ts
@@ -1,3 +1,0 @@
-export type BaseCategories = 'bank' | 'crisis' | 'distribution' | 'governance' | 'slashing' | 'staking' | 'profiles' | 'ibc' | 'ibc-transfer' | 'authz' | 'feegrant' | 'vesting' | 'others'
-export type CustomCategories = ''; // custom modules
-export type Categories = BaseCategories | CustomCategories
